+++ src/platform_impl/linux/x11/util/randr.rs
@@ -126,7 +126,7 @@ impl XConnection {
 
         let scale_factor = match dpi_env {
             EnvVarDPI::Randr => calc_dpi_factor(
-                ((*crtc).width as u32, (*crtc).height as u32),
+                ((*crtc).width, (*crtc).height),
                 (
                     (*output_info).mm_width as u64,
                     (*output_info).mm_height as u64,
@@ -146,7 +146,7 @@ impl XConnection {
                     dpi / 96.
                 } else {
                     calc_dpi_factor(
-                        ((*crtc).width as u32, (*crtc).height as u32),
+                        ((*crtc).width, (*crtc).height),
                         (
                             (*output_info).mm_width as u64,
                             (*output_info).mm_height as u64,
