+++ src/platform_impl/macos/monitor.rs
@@ -200,14 +200,14 @@ impl MonitorHandle {
     pub fn position(&self) -> PhysicalPosition<i32> {
         let bounds = unsafe { CGDisplayBounds(self.native_identifier()) };
         PhysicalPosition::from_logical::<_, f64>(
-            (bounds.origin.x as f64, bounds.origin.y as f64),
+            (bounds.origin.x, bounds.origin.y),
             self.scale_factor(),
         )
     }
 
     pub fn scale_factor(&self) -> f64 {
         match self.ns_screen() {
-            Some(screen) => screen.backingScaleFactor() as f64,
+            Some(screen) => screen.backingScaleFactor(),
             None => 1.0, // default to 1.0 when we can't find the screen
         }
     }
